GUI update 1.1
Added Tolerance box.
Error dialogs when the inputs aren't compatible with the corresponding
box (i.e. typing a string in the tolerance or xi and xf boxes). Still to
do, detect if the function is a valid one.
When pressing the 'calcular' button, it checks which method was
selected, and sends a warning dialog informing which one is checked.
This is just to check if it's working, later the portion of code that
displays the dialog will the replace by the actual method called.